+++ tests/testthat/test_csranks_multinom.R
@@ -8,24 +8,33 @@ test_that("return value is of correct class and size", {
 
     expect_is(res1$L, "integer")
     expect_is(res1$U, "integer")
+    expect_is(res1$rank, "numeric")
     expect_is(res2$L, "integer")
     expect_is(res2$U, "integer")
+    expect_is(res2$rank, "numeric")
     expect_is(res3$L, "integer")
     expect_is(res3$U, "integer")
-
+    expect_is(res3$rank, "numeric")
+    
     expect_equal(length(res1$L), 10)
     expect_equal(length(res1$U), 10)
+    expect_equal(length(res1$rank), 10)
     expect_equal(length(res2$L), 1)
     expect_equal(length(res2$U), 1)
+    expect_equal(length(res2$rank), 1)
     expect_equal(length(res3$L), 3)
     expect_equal(length(res3$U), 3)
-
+    expect_equal(length(res3$rank), 3)
+    
     expect_false(any(is.na(res1$L)))
     expect_false(any(is.na(res1$U)))
+    expect_false(any(is.na(res1$rank)))
     expect_false(any(is.na(res2$L)))
     expect_false(any(is.na(res2$U)))
+    expect_false(any(is.na(res2$rank)))
     expect_false(any(is.na(res3$L)))
     expect_false(any(is.na(res3$U)))
+    expect_false(any(is.na(res3$rank)))
   }
 })
 
@@ -49,9 +58,9 @@ test_that("lower and upper bounds are in the correct range of values", {
       res2 <- csranks_multinom(x, coverage = 0.95, cstype = cstype, multcorr = multcorr, indices = 1)
       res3 <- csranks_multinom(x, coverage = 0.95, cstype = cstype, multcorr = multcorr, indices = 1:3)
 
-      expect_true(all(res1$L <= res1$U))
-      expect_true(all(res2$L <= res2$U))
-      expect_true(all(res3$L <= res3$U))
+      expect_true(all(res1$L <= res1$rank & res1$rank <= res1$U))
+      expect_true(all(res2$L <= res2$rank & res2$rank <= res2$U))
+      expect_true(all(res3$L <= res3$rank & res3$rank <= res3$U))
 
       expect_true(all(res1$L <= 20 & res1$U <= 20 & res1$L >= 1 & res1$U >= 1))
       expect_true(all(res2$L <= 20 & res2$U <= 20 & res2$L >= 1 & res2$U >= 1))